Click here for details
This pdf consists of the Theory behind Logistic Regression and the mathematical proof of how convex optimization can be used to solve the Logistic Regression Problem. It contains the implementations of code where we build a logistic regression model using cvxpy and linearly classify a set of randomly generated points. We also apply this on a real world Dataset concerning with heart disease prediction based on various factors like bp, cholesterol etc.